internal_snapshot handle modified
