+++ internal/snapshotlibvirt/external_snapshot_ops.go
