+++ services/snapshot/internal_snap/delete_int.go
@@ -24,23 +24,14 @@ func deleteSnapshot(domain snapshotDomain, snapName string) error {
 	if err != nil {
 		return virerr.ErrorGen(virerr.SnapshotError, fmt.Errorf("failed to list snapshots: %w", err))
 	}
-	defer func() {
-		for _, s := range snaps {
-			s.Free()
-		}
-	}()
+	defer freeSnapshotHandles(snaps)
 
-	for _, s := range snaps {
-		n, err := s.Name()
-		if err != nil {
-			continue
+	target := findSnapshotByName(snaps, snapName)
+	if target != nil {
+		if err := target.Delete(); err != nil {
+			return virerr.ErrorGen(virerr.SnapshotError, fmt.Errorf("failed to delete snapshot %s: %w", snapName, err))
 		}
-		if n == snapName {
-			if err := s.Delete(); err != nil {
-				return virerr.ErrorGen(virerr.SnapshotError, fmt.Errorf("failed to delete snapshot %s: %w", snapName, err))
-			}
-			return nil
-		}
+		return nil
 	}
 
 	return virerr.ErrorGen(virerr.SnapshotError, fmt.Errorf("snapshot %s not found", snapName))
