+++ services/snapshot/internal_snap/list_int.go
@@ -24,6 +24,7 @@ func listSnapshots(domain snapshotDomain) ([]string, error) {
 	if err != nil {
 		return nil, virerr.ErrorGen(virerr.SnapshotError, fmt.Errorf("failed to list snapshots: %w", err))
 	}
+	defer freeSnapshotHandles(snaps)
 
 	names := make([]string, 0, len(snaps))
 	for _, s := range snaps {
@@ -31,7 +32,6 @@ func listSnapshots(domain snapshotDomain) ([]string, error) {
 		if err == nil {
 			names = append(names, n)
 		}
-		s.Free()
 	}
 
 	return names, nil
